requirements.txt
Flask==2.3.3
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.11.4
google-auth==2.23.4
numpy>=1.24.3
pandas>=1.5.3
textblob==0.17.1
langdetect==1.0.9
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-mock==3.11.1